Change Mutators to act on individual points, not iterators
Requires a rewrite of RandomOffsetMutator, which currently does not
alter the bounds of a point. Doing so would require rethinking mutators
a little bit so the bound information could be updated.
The random offset generation is now required to be deterministic based
on the points passed (since the points can now be generated in a random
order) and must be fully consistent.

Fixes CompoundGenerator to call mutators in get_point.