Change Dimension.create_dimension_mask to use "prepare" pattern
Renames create_dimension_mask to prepare and puts the result into the
dimension.mask attribute. dimension.prepare is now a no-op on repeated
calls and calling get_positions and reading shape/size attributes is not
valid until this method has been called.

This shouldn't matter as it is not expected for Dimension instances to
be created outside CompoundGenerator