Add Jython numpy requirements to Travis build
Instruct Travis to download and extract the requirements from a
release of DAWN. Very heavy handed, but works.

Unfortunately, the exact functionality we require is not neatly
packaged into its own project or module and cannot "neatly" be
extracted, so we end up with a lot of "loose ends". The tests will
run, but a lot of "startup" imports fail and report errors.