Fix coverage support for Jython Travis build
"nosetests --with-coverage" uses coverage.py, which has two phases:

 - The first (run) phase runs the code and records the lines hit.
   This phase *works* when run with Jython.

 - The second (report) phase analyses the source code to determine all
   the possible paths, allowing the "miss" count to be built.
   This phase *fails* when run with Jython.

Nose does not allow omission of the second step, so our Jython tests
will always "fail" with the last error being the failure of the code
coverage report step. We can absorb this error and analyse the test
output ourselves (grep for non-zero Failures/Errors in xunit xml
output), leaving a suitable return code in "$?".

The "python version" of coveralls can then be run and will use the
output of the Jython coverage.py run.