Fix lower/upper bounds for grid special case.
The upper/lower bounds were not being applied to the innermost
line generator in the special case of a pair of line generators
inside an axis-aligned rectangle.

The comparison to detect which generator was the "inner-most" did
not account for the fact that the generators used in the Dimension
structures were not necessarily the same as the as the ones passed
to the CompoundGenerator.