Store fully expanded position arrays in Dimensions
Expand out the position arrays (tile & repeat) for every axis in a
dimension to the full size of the dimension and use these for masking
and indexing.

This simplifies some of the logic to index points and fixes a few
bugs, particularly for bounds calculations, and should facilitate
progress in removing some of the limitations around the "alternate"
setting.

This does greatly increase the memory cost for particularly
"pathological" dimensions (large number of nested generators squashed
by excluders) to the point that they are no longer reasonably
supported. Our intended use cases are far more humble.